Move setup and teardown into a base class
