+++ test/Microsoft.Azure.Devices.Edge.Test/Module.cs
@@ -2,7 +2,6 @@
 namespace Microsoft.Azure.Devices.Edge.Test
 {
     using System;
-    using System.Collections.Generic;
     using System.Threading;
     using System.Threading.Tasks;
     using Microsoft.Azure.Devices.Edge.Test.Common;
@@ -10,45 +9,10 @@ namespace Microsoft.Azure.Devices.Edge.Test
     using Microsoft.Azure.Devices.Edge.Test.Helpers;
     using Microsoft.Azure.Devices.Edge.Util;
     using NUnit.Framework;
-    using NUnit.Framework.Interfaces;
     using Serilog;
 
-    public class Module
+    public class Module : ModuleBase
     {
-        CancellationTokenSource cts;
-        DateTime testStartTime;
-
-        [SetUp]
-        public void BeforeEach()
-        {
-            this.cts = new CancellationTokenSource(Context.Current.TestTimeout);
-            this.testStartTime = DateTime.Now;
-        }
-
-        [TearDown]
-        public async Task AfterEachAsync()
-        {
-            await Profiler.Run(
-                async () =>
-                {
-                    this.cts.Dispose();
-
-                    if (TestContext.CurrentContext.Result.Outcome != ResultState.Ignored)
-                    {
-                        using (var cts = new CancellationTokenSource(Context.Current.TeardownTimeout))
-                        {
-                            string prefix = $"{Context.Current.DeviceId}-{TestContext.CurrentContext.Test.NormalizedName()}";
-                            IEnumerable<string> paths = await EdgeLogs.CollectAsync(this.testStartTime, prefix, cts.Token);
-                            foreach (string path in paths)
-                            {
-                                TestContext.AddTestAttachment(path);
-                            }
-                        }
-                    }
-                },
-                "Completed test teardown");
-        }
-
         [Test]
         public async Task TempSensor()
         {
